LayoutTests/accessibility/gtk/aria-current-changed-notification-expected.txt

This tests that changing the aria-current value results in a state-changed notification.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS item1.currentStateValue is 'false'
PASS item2.currentStateValue is 'page'
PASS item3.currentStateValue is 'false'
Setting aria-current to false on item2.
Setting aria-current to page on item3.
ActiveStateChanged: AXTitle:  AXRole: AXListItem
ActiveStateChanged: AXTitle:  AXRole: AXListItem
PASS item1.currentStateValue is 'false'
PASS item2.currentStateValue is 'false'
PASS item3.currentStateValue is 'page'
PASS successfullyParsed is true

TEST COMPLETE
